fix: restructure event editing functionality and enhance image validation in CreateEventForm
